fix path issues

+++ auth/pages/register.php
@@ -21,20 +21,23 @@ if (isset($_POST['register'])) {
 
     $userMgr = new UserManager();
     if ($userMgr->passwordMatch($password, $confirm_password)) {
+        if ($email != '' && $nome != '' && $cognome != '' && $password != '' && $confirm_password != '' && $street != '' && $city != '' && $cap != '') {
 
-        $result = $userMgr->register($nome, $cognome, $email, $password);
-        if(isset($_POST['newsletter_checked']) && $_POST['newsletter_checked'] === 'on'){
-            $userMgr->addToNewletter($email);
-        }
 
-        if ($result > 0) {
-            $userMgr->createAddress($result, $street, $city, $cap);
-            $userMgr->initializePoints($result);
+            $result = $userMgr->register($nome, $cognome, $email, $password);
+            if (isset($_POST['newsletter_checked']) && $_POST['newsletter_checked'] === 'on') {
+                $userMgr->addToNewletter($email);
+            }
 
-            echo '<script>location.href="' . ROOT_URL . 'auth?page=login"</script>';
-            exit;
-        } else {
-            $errMsg = 'Mail già utilizzata...';
+            if ($result > 0) {
+                $userMgr->createAddress($result, $street, $city, $cap);
+                $userMgr->initializePoints($result);
+
+                echo '<script>location.href="' . ROOT_URL . 'auth?page=login"</script>';
+                exit;
+            } else {
+                $errMsg = 'Mail già utilizzata...';
+            }
         }
     } else {
         $errMsg = "Le password non corrsipondono...";
